add wikiextractor to clean text

+++ main.py
@@ -52,7 +52,7 @@ def retrieveData(file_name_pattern, file_path = 'data/compressed/', baseURL = "h
 				for data in tqdm(response.iter_content(block_size), total=math.ceil(file_size//block_size) , unit='KB', unit_scale=True):
 					compressed_file.write(data)
 
-def uncompressData(input_path = 'data/compressed/', output_path = 'data/unprocessed/'):
+def uncompressData(input_path = 'data/compressed/', output_path = 'data/unprocessed/', remove_file = False):
 	# Preprocess the total number of files
 	filetotal = 0
 	for file in iter_files(input_path):
@@ -73,13 +73,14 @@ def uncompressData(input_path = 'data/compressed/', output_path = 'data/unproces
 						#Decompress file and save
 						buf = gzip.decompress(compressed_file.read())
 						uncompressed_file.write(buf)
-			#Delete comesseprd file after uncompressing
-			os.remove(file)
+			#Delete compressed file after uncompressing
+			if remove_file:
+				os.remove(file)
 	logger.info('Finish extracting %d files.' %filetotal)
 
 
 #Retrieve content from xml files
-def processData(input_path = 'data/compressed/', output_path = 'data/unprocessed/', remove_formatting = False):
+def processData(input_path = 'data/compressed/', output_path = 'data/unprocessed/', remove_formatting = False, remove_file = False):
 	#global dictionary of documents + revision numbers vs categories
 	category_files = defaultdict(dict)
 	#local dict to store all revisions of category of each document
@@ -104,49 +105,51 @@ def processData(input_path = 'data/compressed/', output_path = 'data/unprocessed
 			file_name = os.path.basename(file)
 			logger.info('Processing %d/%d files...' %(filecounter, filetotal))
 			tree = ET.parse(file)
+			
 			#Get root tag <page>
 			root = tree.getroot()[1]
 			file_start = '{http://www.mediawiki.org/xml/export-0.10/}'
 
-			#Get text content and revision number
-			revision_count = 1
-			data = {}
-			for revision in root.iter(file_start+'revision'):
-				text = revision.find(file_start+'text').text
-				if text != None:
-					#category of each revision
-					categories = category_content_pattern.findall(text)
-					if len(categories):  #if there is any categories
-						for cat in categories:  #for each category
-							if cat not in local_dict:  #if category hasn't existed
-							# 	local_dict[cat] = []	#create an array to store revision numbers
-							# else:						#uncomment this to have array of all revision numbers
-							# 	local_dict[cat].append(i)
-								local_dict[cat] = revision_count
+				
+			json_file_name = file_name[:-3] + 'json'
+			with open(output_path+json_file_name, 'w') as json_file:
+				revision_count = 0
+				data = {}
+				for revision in root.iter(file_start+'revision'):
+					revision_count += 1
+					text = revision.find(file_start+'text').text
+					if text != None:
+						#retrieve categories and references from text
+						categories = category_content_pattern.findall(text)
+						if len(categories): 
+							for cat in categories:  
+								if cat not in local_dict:  
+									local_dict[cat] = revision_count
 
-					if remove_formatting:	
-						text, refs = stripFormatting(text)
-					data[i] = tokenization(text), categories, refs
-				revision_count += 1
+						if remove_formatting:	
+							text, refs = stripFormatting(text)
+						data[revision_count] = {
+							'text': text,
+							'categories': categories,
+							'references': refs
+						}
+				#Write to json file
+				json.dump(data, json_file, indent = 2)
 
-			for category, revision_number in local_dict.items():
-				if category in category_files:  #if a category file has existed
-					with open('data/categories/'+category+".txt", "a") as category_file:  #append to the file
-						#store document's name and revision number
-						category_file.write(file_name + ' ' + str(revision_number) + '\n')
-						# for num in revision_number:	#uncomment this to have array of all revision numbers
-						# 	category_file.write(' '+ str(num))
-				else:
-					with open('data/categories/'+category+".txt", "w") as category_file:  #create a new file
-						category_file.write(file_name + ' ' + str(revision_number) + '\n')
-						# for num in revision_number:
-						# 	category_file.write(' '+ str(num))
-
-				#Write a new file with filtered data as json format
-				json_file_name = file_name[:-3] + 'json'
-				with open(output_path+json_file_name, 'w') as data_file:
-					json.dump(data, data_file, indent = 2)
-				#Delete uncompressed after finish processing
+				# for category, revision_number in local_dict.items():
+				# 	if category in category_files:  #if a category file has existed
+				# 		with open('data/categories/'+category+".txt", "a") as category_file:  #append to the file
+				# 			#store document's name and revision number
+				# 			category_file.write(file_name + ' ' + str(revision_number) + '\n')
+				# 			# for num in revision_number:	#uncomment this to have array of all revision numbers
+				# 			# 	category_file.write(' '+ str(num))
+				# 	else:
+				# 		with open('data/categories/'+category+".txt", "w") as category_file:  #create a new file
+				# 			category_file.write(file_name + ' ' + str(revision_number) + '\n')
+				# 			# for num in revision_number:
+				# 			# 	category_file.write(' '+ str(num))
+			#Delete original file
+			if remove_file:
 				os.remove(file)
 	logger.info('Finish processing %d files.' %filetotal)
 
@@ -199,8 +202,8 @@ if __name__ == '__main__':
 	# what we need: 'enwiki-latest-stub-meta-history[0-9]{0,3}.xml.gz'
 	file_pattern = re.compile('enwiki-latest-abstract11.xml.gz')
 	
-	retrieveData(file_pattern)
-	uncompressData()
+	# retrieveData(file_pattern)
+	# uncompressData()
 	processData(remove_formatting = True)
 	# call php script
 	result = subprocess.run(
